+++ scripts/build_kraken2_db.sh
@@ -99,6 +99,9 @@ echo "Estimating required capacity (step 2)..."
 
 step_time=$(get_current_time)
 estimate=$(list_sequence_files | xargs -0 cat | estimate_capacity -k $KRAKEN2_KMER_LEN -l $KRAKEN2_MINIMIZER_LEN -S $KRAKEN2_SEED_TEMPLATE -p $KRAKEN2_THREAD_CT $KRAKEN2XFLAG )
+# Slight upward adjustment of distinct minimizer estimate to protect
+# against crash w/ small reference sets
+estimate=$(( estimate + 8192 ))
 required_capacity=$(perl -le 'print int(shift() / shift())' $estimate $KRAKEN2_LOAD_FACTOR);
 
 echo "Estimated hash table requirement: $(( required_capacity * 4 )) bytes"
@@ -117,6 +120,12 @@ echo "Capacity estimation complete. [$(report_time_elapsed $step_time)]"
 
 echo "Building database files (step 3)..."
 
+fast_build_flag=""
+if [ -n "$KRAKEN2_FAST_BUILD" ]
+then
+  fast_build_flag="-F"
+fi
+
 if [ -e "hash.k2d" ]
 then
   echo "Hash table already present, skipping database file build."
@@ -125,7 +134,8 @@ else
   list_sequence_files | xargs -0 cat | \
     build_db -k $KRAKEN2_KMER_LEN -l $KRAKEN2_MINIMIZER_LEN -S $KRAKEN2_SEED_TEMPLATE $KRAKEN2XFLAG \
              -H hash.k2d.tmp -t taxo.k2d.tmp -o opts.k2d.tmp -n taxonomy/ -m $seqid2taxid_map_file \
-             -c $required_capacity -p $KRAKEN2_THREAD_CT $max_db_flag
+             -c $required_capacity -p $KRAKEN2_THREAD_CT $max_db_flag -B $KRAKEN2_BLOCK_SIZE -b $KRAKEN2_SUBBLOCK_SIZE \
+             -r $KRAKEN2_MIN_TAXID_BITS $fast_build_flag
   finalize_file taxo.k2d
   finalize_file opts.k2d
   finalize_file hash.k2d