+++ src/classify.cc
@@ -14,6 +14,8 @@
 #include "aa_translate.h"
 #include "reports.h"
 #include "utilities.h"
+#include "readcounts.h"
+using namespace kraken2;
 
 using std::cout;
 using std::cerr;
@@ -41,6 +43,7 @@ struct Options {
   string unclassified_output_filename;
   string kraken_output_filename;
   bool mpa_style_report;
+  bool report_kmer_data;
   bool quick_mode;
   bool report_zero_counts;
   bool use_translated_search;
@@ -52,6 +55,7 @@ struct Options {
   int minimum_quality_score;
   int minimum_hit_groups;
   bool use_memory_mapping;
+  bool match_input_order;
 };
 
 struct ClassificationStats {
@@ -84,12 +88,12 @@ void usage(int exit_code=EX_USAGE);
 void ProcessFiles(const char *filename1, const char *filename2,
     KeyValueStore *hash, Taxonomy &tax,
     IndexOptions &idx_opts, Options &opts, ClassificationStats &stats,
-    OutputStreamData &outputs, taxon_counts_t &call_counts);
+    OutputStreamData &outputs, taxon_counters_t &total_taxon_counters);
 taxid_t ClassifySequence(Sequence &dna, Sequence &dna2, ostringstream &koss,
     KeyValueStore *hash, Taxonomy &tax, IndexOptions &idx_opts,
     Options &opts, ClassificationStats &stats, MinimizerScanner &scanner,
     vector<taxid_t> &taxa, taxon_counts_t &hit_counts,
-    vector<string> &tx_frames);
+    vector<string> &tx_frames, taxon_counters_t &my_taxon_counts);
 void AddHitlistString(ostringstream &oss, vector<taxid_t> &taxa,
     Taxonomy &taxonomy);
 taxid_t ResolveTree(taxon_counts_t &hit_counts,
@@ -107,6 +111,7 @@ int main(int argc, char **argv) {
   opts.single_file_pairs = false;
   opts.num_threads = 1;
   opts.mpa_style_report = false;
+  opts.report_kmer_data = false;
   opts.report_zero_counts = false;
   opts.use_translated_search = false;
   opts.print_scientific_name = false;
@@ -114,6 +119,7 @@ int main(int argc, char **argv) {
   opts.minimum_hit_groups = 0;
   opts.use_memory_mapping = false;
 
+  taxon_counters_t taxon_counters; // stats per taxon
   ParseCommandLine(argc, argv, opts);
 
   omp_set_num_threads(opts.num_threads);
@@ -122,7 +128,11 @@ int main(int argc, char **argv) {
 
   IndexOptions idx_opts = {0};
   ifstream idx_opt_fs(opts.options_filename);
-  idx_opt_fs.read((char *) &idx_opts, sizeof(idx_opts));
+  struct stat sb;
+  if (stat(opts.options_filename.c_str(), &sb) < 0)
+    errx(EX_OSERR, "unable to get filesize of %s", opts.options_filename.c_str());
+  auto opts_filesize = sb.st_size;
+  idx_opt_fs.read((char *) &idx_opts, opts_filesize);
   opts.use_translated_search = ! idx_opts.dna_db;
 
   Taxonomy taxonomy(opts.taxonomy_filename, opts.use_memory_mapping);
@@ -131,7 +141,6 @@ int main(int argc, char **argv) {
   cerr << " done." << endl;
 
   ClassificationStats stats = {0, 0, 0};
-  taxon_counts_t call_counts;
 
   OutputStreamData outputs = { false, false, nullptr, nullptr, nullptr, nullptr, &std::cout };
 
@@ -140,7 +149,7 @@ int main(int argc, char **argv) {
   if (optind == argc) {
     if (opts.paired_end_processing && ! opts.single_file_pairs)
       errx(EX_USAGE, "paired end processing used with no files specified");
-    ProcessFiles(nullptr, nullptr, hash_ptr, taxonomy, idx_opts, opts, stats, outputs, call_counts);
+    ProcessFiles(nullptr, nullptr, hash_ptr, taxonomy, idx_opts, opts, stats, outputs, taxon_counters);
   }
   else {
     for (int i = optind; i < argc; i++) {
@@ -148,11 +157,11 @@ int main(int argc, char **argv) {
         if (i + 1 == argc) {
           errx(EX_USAGE, "paired end processing used with unpaired file");
         }
-        ProcessFiles(argv[i], argv[i+1], hash_ptr, taxonomy, idx_opts, opts, stats, outputs, call_counts);
+        ProcessFiles(argv[i], argv[i+1], hash_ptr, taxonomy, idx_opts, opts, stats, outputs, taxon_counters);
         i += 1;
       }
       else {
-        ProcessFiles(argv[i], nullptr, hash_ptr, taxonomy, idx_opts, opts, stats, outputs, call_counts);
+        ProcessFiles(argv[i], nullptr, hash_ptr, taxonomy, idx_opts, opts, stats, outputs, taxon_counters);
       }
     }
   }
@@ -165,11 +174,12 @@ int main(int argc, char **argv) {
   if (! opts.report_filename.empty()) {
     if (opts.mpa_style_report)
       ReportMpaStyle(opts.report_filename, opts.report_zero_counts, taxonomy,
-          call_counts);
+          taxon_counters);
     else {
       auto total_unclassified = stats.total_sequences - stats.total_classified;
-      ReportKrakenStyle(opts.report_filename, opts.report_zero_counts, taxonomy,
-          call_counts, stats.total_sequences, total_unclassified);
+      ReportKrakenStyle(opts.report_filename, opts.report_zero_counts,
+          opts.report_kmer_data, taxonomy,
+          taxon_counters, stats.total_sequences, total_unclassified);
     }
   }
 
@@ -211,7 +221,8 @@ void ReportStats(struct timeval time1, struct timeval time2,
 void ProcessFiles(const char *filename1, const char *filename2,
     KeyValueStore *hash, Taxonomy &tax,
     IndexOptions &idx_opts, Options &opts, ClassificationStats &stats,
-    OutputStreamData &outputs, taxon_counts_t &call_counts)
+    OutputStreamData &outputs,
+    taxon_counters_t &total_taxon_counters)
 {
   std::istream *fptr1 = nullptr, *fptr2 = nullptr;
 
@@ -245,12 +256,12 @@ void ProcessFiles(const char *filename1, const char *filename2,
     taxon_counts_t hit_counts;
     ostringstream kraken_oss, c1_oss, c2_oss, u1_oss, u2_oss;
     ClassificationStats thread_stats = {0, 0, 0};
-    vector<taxid_t> calls;
     vector<string> translated_frames(6);
     BatchSequenceReader reader1, reader2;
     Sequence seq1, seq2;
     uint64_t block_id;
     OutputData out_data;
+    taxon_counters_t thread_taxon_counters;
 
     while (true) {
       thread_stats.total_sequences = 0;
@@ -285,12 +296,12 @@ void ProcessFiles(const char *filename1, const char *filename2,
         break;
 
       // Reset all dynamically-growing things
-      calls.clear();
       kraken_oss.str("");
       c1_oss.str("");
       c2_oss.str("");
       u1_oss.str("");
       u2_oss.str("");
+      thread_taxon_counters.clear();
 
       while (true) {
         auto valid_fragment = reader1.NextSequence(seq1);
@@ -310,23 +321,23 @@ void ProcessFiles(const char *filename1, const char *filename2,
         }
         auto call = ClassifySequence(seq1, seq2,
             kraken_oss, hash, tax, idx_opts, opts, thread_stats, scanner,
-            taxa, hit_counts, translated_frames);
+            taxa, hit_counts, translated_frames, thread_taxon_counters);
         if (call) {
           char buffer[1024] = "";
           sprintf(buffer, " kraken:taxid|%llu",
-			  (unsigned long long) tax.nodes()[call].external_id);
+              (unsigned long long) tax.nodes()[call].external_id);
           seq1.header += buffer;
-          seq2.header += buffer;
           c1_oss << seq1.to_string();
-          if (opts.paired_end_processing)
+          if (opts.paired_end_processing) {
+            seq2.header += buffer;
             c2_oss << seq2.to_string();
+          }
         }
         else {
           u1_oss << seq1.to_string();
           if (opts.paired_end_processing)
             u2_oss << seq2.to_string();
         }
-        calls.push_back(call);
         thread_stats.total_bases += seq1.seq.size();
         if (opts.paired_end_processing)
           thread_stats.total_bases += seq2.seq.size();
@@ -346,13 +357,6 @@ void ProcessFiles(const char *filename1, const char *filename2,
                << " sequences (" << stats.total_bases << " bp) ...";
       }
 
-      #pragma omp critical(update_calls)
-      {
-        for (auto &call : calls) {
-          call_counts[call]++;
-        }
-      }
-
       if (! outputs.initialized) {
         InitializeOutputs(opts, outputs, reader1.file_format());
       }
@@ -367,6 +371,13 @@ void ProcessFiles(const char *filename1, const char *filename2,
       #pragma omp critical(output_queue)
       {
         output_queue.push(out_data);
+      }
+
+      #pragma omp critical(update_taxon_counters)
+      {
+        for (auto &kv_pair : thread_taxon_counters) {
+          total_taxon_counters[kv_pair.first] += std::move(kv_pair.second);
+        }
       }
 
       bool output_loop = true;
@@ -389,8 +400,6 @@ void ProcessFiles(const char *filename1, const char *filename2,
         }
         if (! output_loop)
           break;
-        // Past this point in loop, we know lock is set
-
         if (outputs.kraken_output != nullptr)
           (*outputs.kraken_output) << out_data.kraken_str;
         if (outputs.classified_output1 != nullptr)
@@ -489,7 +498,8 @@ taxid_t ClassifySequence(Sequence &dna, Sequence &dna2, ostringstream &koss,
     KeyValueStore *hash, Taxonomy &taxonomy, IndexOptions &idx_opts,
     Options &opts, ClassificationStats &stats, MinimizerScanner &scanner,
     vector<taxid_t> &taxa, taxon_counts_t &hit_counts,
-    vector<string> &tx_frames)
+    vector<string> &tx_frames,
+    taxon_counters_t &curr_taxon_counts)
 {
   uint64_t *minimizer_ptr;
   taxid_t call = 0;
@@ -515,10 +525,9 @@ taxid_t ClassifySequence(Sequence &dna, Sequence &dna2, ostringstream &koss,
       }
       uint64_t last_minimizer = UINT64_MAX;
       taxid_t last_taxon = TAXID_MAX;
-      bool ambig_flag = false;
-      while ((minimizer_ptr = scanner.NextMinimizer(&ambig_flag)) != nullptr) {
+      while ((minimizer_ptr = scanner.NextMinimizer()) != nullptr) {
         taxid_t taxon;
-        if (ambig_flag) {
+        if (scanner.is_ambiguous()) {
           taxon = AMBIGUOUS_SPAN_TAXON;
         }
         else {
@@ -535,8 +544,11 @@ taxid_t ClassifySequence(Sequence &dna, Sequence &dna2, ostringstream &koss,
             last_minimizer = *minimizer_ptr;
             // Increment this only if (a) we have DB hit and
             // (b) minimizer != last minimizer
-            if (taxon)
+            if (taxon) {
               minimizer_hit_groups++;
+              // New minimizer should trigger registering minimizer in RC/HLL
+              curr_taxon_counts[taxon].add_kmer(scanner.last_minimizer());
+            }
           }
           else {
             taxon = last_taxon;
@@ -570,8 +582,10 @@ taxid_t ClassifySequence(Sequence &dna, Sequence &dna2, ostringstream &koss,
   if (call && minimizer_hit_groups < opts.minimum_hit_groups)
     call = 0;
 
-  if (call)
+  if (call) {
     stats.total_classified++;
+    curr_taxon_counts[call].incrementReadCount();
+  }
 
   if (call)
     koss << "C\t";
@@ -601,7 +615,7 @@ taxid_t ClassifySequence(Sequence &dna, Sequence &dna2, ostringstream &koss,
     koss << dna.seq.size() << "|" << dna2.seq.size() << "\t";
 
   if (opts.quick_mode) {
-    koss << call << ":Q";
+    koss << ext_call << ":Q";
   }
   else {
     if (taxa.empty())
@@ -724,7 +738,7 @@ void MaskLowQualityBases(Sequence &dna, int minimum_quality_score) {
 void ParseCommandLine(int argc, char **argv, Options &opts) {
   int opt;
 
-  while ((opt = getopt(argc, argv, "h?H:t:o:T:p:R:C:U:O:Q:g:nmzqPSM")) != -1) {
+  while ((opt = getopt(argc, argv, "h?H:t:o:T:p:R:C:U:O:Q:g:nmzqPSMK")) != -1) {
     switch (opt) {
       case 'h' : case '?' :
         usage(0);
@@ -764,6 +778,9 @@ void ParseCommandLine(int argc, char **argv, Options &opts) {
         break;
       case 'm' :
         opts.mpa_style_report = true;
+        break;
+      case 'K':
+        opts.report_kmer_data = true;
         break;
       case 'R' :
         opts.report_filename = optarg;
@@ -827,6 +844,7 @@ void usage(int exit_code) {
        << "  -g NUM           Minimum number of hit groups needed for call" << endl
        << "  -C filename      Filename/format to have classified sequences" << endl
        << "  -U filename      Filename/format to have unclassified sequences" << endl
-       << "  -O filename      Output file for normal Kraken output" << endl;
+       << "  -O filename      Output file for normal Kraken output" << endl
+       << "  -K               In comb. w/ -R, provide minimizer information in report" << endl;
   exit(exit_code);
 }