+++ src/reports.cc
@@ -29,6 +29,22 @@ taxon_counts_t GetCladeCounts(Taxonomy &tax, taxon_counts_t &call_counts) {
   return clade_counts;
 }
 
+taxon_counters_t GetCladeCounters(Taxonomy &tax, taxon_counters_t &call_counters) {
+  taxon_counters_t clade_counters;
+
+  for (auto &kv_pair : call_counters) {
+    auto taxid = kv_pair.first;
+    auto counter = kv_pair.second;
+
+    while (taxid) {
+      clade_counters[taxid] += counter;
+      taxid = tax.nodes()[taxid].parent_id;
+    }
+  }
+
+  return clade_counters;
+}
+
 void PrintMpaStyleReportLine(ofstream &ofs, uint64_t clade_count, const string &taxonomy_line) {
   ofs << taxonomy_line << "\t" << clade_count << std::endl;
 }
@@ -88,23 +104,33 @@ void MpaReportDFS(taxid_t taxid, ofstream &ofs, bool report_zeros, Taxonomy &tax
     taxonomy_names.pop_back();
 }
 
-void ReportMpaStyle(string filename, bool report_zeros, Taxonomy &taxonomy, taxon_counts_t &call_counts) {
+void ReportMpaStyle(string filename, bool report_zeros, Taxonomy &taxonomy, taxon_counters_t &call_counters) {
+  taxon_counts_t call_counts;
+  for (auto &kv_pair : call_counters) {
+    call_counts[kv_pair.first] = kv_pair.second.readCount();
+  }
   taxon_counts_t clade_counts = GetCladeCounts(taxonomy, call_counts);
   ofstream ofs(filename);
   vector<string> taxonomy_names;
   MpaReportDFS(1, ofs, report_zeros, taxonomy, clade_counts, taxonomy_names);
 }
 
-void PrintKrakenStyleReportLine(ofstream &ofs, uint64_t total_seqs, uint64_t clade_count, uint64_t call_count,
+void PrintKrakenStyleReportLine(ofstream &ofs, bool report_kmer_data,
+  uint64_t total_seqs,
+  READCOUNTER clade_counter, READCOUNTER taxon_counter,
   const string &rank_str, uint32_t taxid, const string &sci_name, int depth)
 {
   char pct_buffer[10] = "";
-  sprintf(pct_buffer, "%6.2f", 100.0 * clade_count / total_seqs);
+  sprintf(pct_buffer, "%6.2f", 100.0 * clade_counter.readCount() / total_seqs);
 
   ofs << pct_buffer << "\t"
-      << clade_count << "\t"
-      << call_count << "\t"
-      << rank_str << "\t"
+      << clade_counter.readCount() << "\t"
+      << taxon_counter.readCount() << "\t";
+  if (report_kmer_data) {
+    ofs << clade_counter.kmerCount() << "\t"
+        << clade_counter.distinctKmerCount() << "\t";
+  }
+  ofs << rank_str << "\t"
       << taxid << "\t";
   for (auto i = 0; i < depth; i++)
     ofs << "  ";
@@ -113,12 +139,13 @@ void PrintKrakenStyleReportLine(ofstream &ofs, uint64_t total_seqs, uint64_t cla
 
 // Depth-first search of taxonomy tree, reporting info at each node
 void KrakenReportDFS(uint32_t taxid, ofstream &ofs, bool report_zeros,
-    Taxonomy &taxonomy, taxon_counts_t &clade_counts,
-    taxon_counts_t &call_counts, uint64_t total_seqs,
+    bool report_kmer_data,
+    Taxonomy &taxonomy, taxon_counters_t &clade_counters,
+    taxon_counters_t &call_counters, uint64_t total_seqs,
     char rank_code, int rank_depth, int depth)
 {
   // Clade count of 0 means all subtree nodes have clade count of 0
-  if (! report_zeros && clade_counts[taxid] == 0)
+  if (! report_zeros && clade_counters[taxid].readCount() == 0)
     return;
   TaxonomyNode node = taxonomy.nodes()[taxid];
   string rank = taxonomy.rank_data() + node.rank_offset;
@@ -139,8 +166,9 @@ void KrakenReportDFS(uint32_t taxid, ofstream &ofs, bool report_zeros,
 
   string name = taxonomy.name_data() + node.name_offset;
 
-  PrintKrakenStyleReportLine(ofs, total_seqs, clade_counts[taxid],
-      call_counts[taxid], rank_str, node.external_id, name, depth);
+  PrintKrakenStyleReportLine(ofs, report_kmer_data, total_seqs,
+    clade_counters[taxid], call_counters[taxid], rank_str, node.external_id,
+    name, depth);
 
   auto child_count = node.child_count;
   if (child_count != 0) {
@@ -149,33 +177,37 @@ void KrakenReportDFS(uint32_t taxid, ofstream &ofs, bool report_zeros,
     for (auto i = 0u; i < child_count; i++) {
       children[i] = node.first_child + i;
     }
-    // Sorting child IDs by descending order of clade counts
+    // Sorting child IDs by descending order of clade read counts
     std::sort(children.begin(), children.end(),
       [&](const uint64_t &a, const uint64_t &b) {
-        return clade_counts[a] > clade_counts[b];
+        return clade_counters[a].readCount() > clade_counters[b].readCount();
       }
     );
     for (auto child : children) {
-      KrakenReportDFS(child, ofs, report_zeros, taxonomy, clade_counts, call_counts,
-          total_seqs, rank_code, rank_depth, depth + 1);
+      KrakenReportDFS(child, ofs, report_zeros, report_kmer_data, taxonomy,
+          clade_counters, call_counters, total_seqs, rank_code, rank_depth,
+          depth + 1);
     }
   }
 }
 
-void ReportKrakenStyle(string filename, bool report_zeros, Taxonomy &taxonomy,
-    taxon_counts_t &call_counts, uint64_t total_seqs, uint64_t total_unclassified)
+void ReportKrakenStyle(string filename, bool report_zeros, bool report_kmer_data,
+    Taxonomy &taxonomy, taxon_counters_t &call_counters, uint64_t total_seqs,
+    uint64_t total_unclassified)
 {
-  taxon_counts_t clade_counts = GetCladeCounts(taxonomy, call_counts);
+  taxon_counters_t clade_counters = GetCladeCounters(taxonomy, call_counters);
 
   ofstream ofs(filename);
 
   // Special handling of the unclassified sequences
-  if (total_unclassified != 0 || report_zeros)
-    PrintKrakenStyleReportLine(ofs, total_seqs, total_unclassified,
-                               total_unclassified, "U", 0, "unclassified", 0);
+  if (total_unclassified != 0 || report_zeros) {
+    READCOUNTER rc(total_unclassified, 0);
+    PrintKrakenStyleReportLine(ofs, report_kmer_data, total_seqs, rc,
+                               rc, "U", 0, "unclassified", 0);
+  }
   // DFS through the taxonomy, printing nodes as encountered
-  KrakenReportDFS(1, ofs, report_zeros, taxonomy, clade_counts, call_counts,
-                  total_seqs, 'R', -1, 0);
+  KrakenReportDFS(1, ofs, report_zeros, report_kmer_data, taxonomy,
+                  clade_counters, call_counters, total_seqs, 'R', -1, 0);
 }
 
 }  // end namespace